+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 TBD
 
+- Added new beer.conf support for shutdown functions. These functions will be executed at the start
+  of the shutdown process. Function must be present on the client class object.
+- Added new beer.conf support for startup functions. These functions will be executed after `Plugin().run()` 
+  has completed startup processes. Function must be present on the client class object.
 - Added python version to about page
 - Added Rocky 9 RPM Support